MatrixPilot_beta: sync my branch and add revamped build system, with STM32 support. next step is to test a fresh checkout under linux.
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_beta@3488 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ libDCM/libDCM_defines.h
@@ -95,7 +95,8 @@ struct dcm_flag_bits {
 
 #define ACCELSCALE ((int32_t)(GRAVITY/GRAVITYM))
 
-#define CENTRISCALE ((int32_t)(((int64_t)519168.0)*GRAVITY)/((int64_t)RADPERSEC*GRAVITYM))
+//#define CENTRISCALE ((int32_t)(((int64_t)519168.0)*GRAVITY)/((int64_t)RADPERSEC*GRAVITYM))
+#define CENTRISCALE ((uint64_t)((((uint64_t)519168.0)*(uint64_t)GRAVITY)/((uint64_t)RADPERSEC*(uint64_t)GRAVITYM)))
 // scale factor in multiplying omega times velocity to get centrifugal acceleration
 
 #define CENTRIFSAT ((int32_t)(GRAVITYM*RADPERSEC)/(GRAVITY*((int64_t)32)))
